requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
filelock==3.12.0
fonttools==4.40.0
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.30.0
six==1.16.0
sympy==1.12
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2
typing_extensions==4.5.0
urllib3==2.0.2
pyqt5
